backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dateutil==2.8.2
bleach==6.0.0
sqlalchemy==2.0.23